+++ launcher-core/src/main/java/com/launcher/core/operation/type/OperationType.java
@@ -4,5 +4,5 @@ public enum OperationType {
 
     REPAIR,
     LOAD_MANIFEST,
-
+    VERIFY_FILES
 }